deployment-models/requirements.txt
python-dotenv==0.13.0
mlflow==1.8.0
boto3==1.13.12